- avoid dependency on Laravel collections - ability to specify a job name rather than a stage name - fix GitLab API URLs - optionally preserve tarballs - don't fail if storage/ does not exist

+++ src/Console/Gitlab/ArtifactDownload.php
@@ -28,7 +28,7 @@ class Download extends Command
     /**
      * Gitlab API URL to download artifact for a given build
      */
-    const ARTIFACT_BUILD_URL = "%s/api/v4/projects/%s/artifacts/%d/artifacts/%s/download?job=%s";
+    const ARTIFACT_BUILD_URL = "%s/api/v4/projects/%s/jobs/%d/artifacts";
 
     /**
      * The name and signature of the console command.
@@ -61,7 +61,9 @@ class Download extends Command
             ->addOption('in', null, InputOption::VALUE_REQUIRED, 'Path where the artifact must be extracted', '.')
             ->addOption('ref', null, InputOption::VALUE_REQUIRED, 'Repository ref the build was ran on')
             ->addOption('tag', null, InputOption::VALUE_REQUIRED, 'Repository tag the build was ran on')
-            ->addOption('stage', null, InputOption::VALUE_REQUIRED, 'Build stage from which artifact is downloaded', 'prepare');
+            ->addOption('stage', null, InputOption::VALUE_REQUIRED, 'Build stage from which artifact is downloaded')
+            ->addOption('name', null, InputOption::VALUE_REQUIRED, 'Job name from which artifact', 'prepare')
+            ->addOption('preserve-temp-files', null, InputOption::VALUE_NONE, 'Whether to preserve downloaded tarball');
     }
 
     public function __construct($composer, $io) {
@@ -102,6 +104,9 @@ class Download extends Command
         if (null === $token = $this->option('token')) {
             $token = env('GITLAB_TOKEN');
         }
+        if (!$this->option('name') && !$this->option('stage')) {
+            throw new \RuntimeException('At least one of --name or --stage must be supplied.');
+        }
 
         $in = realpath($this->option('in'));
         if (!is_dir($in) || !is_writable($in)) {
@@ -116,13 +121,17 @@ class Download extends Command
             $token,
             $project,
             $this->option('stage'),
+            $this->option('name'),
             $this->option('ref'),
             $this->option('tag')
         );
 
         $path = base_path().'/storage/artifact-'.$build['id'].'.zip';
         try {
-            $zip = fopen($path, 'w+');
+            if (! is_dir(base_path().'/storage')) {
+                mkdir(base_path().'/storage', 0700);
+            }
+            $zip = fopen($path, 'wb');
             $this->api(
                 sprintf(self::ARTIFACT_BUILD_URL, $url, urlencode($project), $build['id']),
                 $token,
@@ -131,11 +140,16 @@ class Download extends Command
             fclose($zip);
         } catch (\Exception $error) {
             $this->error('Can\'t download artifact to '.$path);
+            $this->error($error->getMessage());
             fclose($zip);
+            return;
         }
+        $this->info(sprintf("Downloaded artifacts tarball at %s (%d bytes)", $path, filesize($path)));
 
         $this->installArtifact($path, $in);
-        unlink($path);
+        if (! $this->option('preserve-temp-files')) {
+            unlink($path);
+        }
     }
 
     /**
@@ -146,28 +160,30 @@ class Download extends Command
      * @return array
      * @throws \RuntimeException
      */
-    private function getLatestSuccessfulBuild($url, $token, $project, $stage, $ref, $tag)
+    private function getLatestSuccessfulBuild($url, $token, $project, $stage, $name, $ref, $tag)
     {
         $result = $this->apiJson(
             sprintf(self::PROJECT_BUILD_URL, $url, urlencode($project)),
             $token
         );
-
-        $result = collect($result)->reject(function ($item) use ($stage, $ref, $tag) {
-            return $item['status'] !== 'success'
-                || $item['stage'] !== $stage
-                || (null !== $ref && $item['ref'] !== $ref)
-                || (null !== $tag && $item['tag'] !== $tag);
-        })->first();
-        if (null === $result) {
+        $result = array_filter($result, function ($item) use ($stage, $name, $ref, $tag) {
+            return $item['status'] === 'success'
+                                   && (!$stage || $item['stage'] === $stage)
+                                   && (!$name || $item['name'] === $name)
+                                   && (!$ref || $item['ref'] === $ref)
+                                   && (!$tag || $item['tag'] === $tag);
+        });
+        $result = reset($result);
+        if (!$result) {
             throw new \RuntimeException('Can\'t find a successful build in the project...');
         }
 
         $this->info(sprintf(
-            "Latest build [%s]\n- ref: %s\n- stage: %s\n- at: %s\n- runner: %d -> %s\n- triggered by: %s",
+            "Latest build [%s]\n- ref: %s\n- stage: %s\n- name: %s\n- at: %s\n- runner: %d -> %s\n- triggered by: %s",
             $result['id'],
             $result['ref'],
             $result['stage'],
+            $result['name'],
             $result['created_at'],
             $result['runner']['id'],
             $result['runner']['description'],
@@ -194,6 +210,7 @@ class Download extends Command
             curl_setopt($h, CURLOPT_FILE, $file);
         }
         curl_setopt($h, CURLOPT_URL, $url);
+        curl_setopt($h, CURLOPT_FOLLOWLOCATION, TRUE);
         curl_setopt($h, CURLOPT_HTTPHEADER, [
             sprintf("PRIVATE-TOKEN: %s", $token)
         ]);
